scmi: Use separate modules for device implementations
For a better readability once we start adding more devices.
