+++ crates/scmi/src/devices/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-FileCopyrightText: Red Hat, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod common;
+pub mod fake;
